finished route setup and testing
